Added response entities for  post, put and delete endpoints

+++ src/main/java/com/brk/commentor/controller/CommentController.java
@@ -4,10 +4,14 @@ import com.brk.commentor.config.CommentModelAssembler;
 import com.brk.commentor.exception.CommentNotFoundException;
 import com.brk.commentor.model.Comment;
 import com.brk.commentor.repository.CommentRepository;
+import org.hibernate.boot.model.source.spi.CascadeStyleSource;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
+import org.springframework.hateoas.IanaLinkRelations;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
@@ -34,8 +38,9 @@ public class CommentController {
     }
 
     @PostMapping("/comments")
-    public Comment newComment(@RequestBody Comment newComment) {
-        return repository.save(newComment);
+    public ResponseEntity<?> newComment(@RequestBody Comment newComment) {
+        EntityModel<Comment> entityModel = assembler.toModel(repository.save(newComment));
+        return ResponseEntity.created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()).body(entityModel);
     }
 
     @GetMapping("/comments/{id}")
@@ -46,19 +51,28 @@ public class CommentController {
     }
 
     @PutMapping("/comments/{id}")
-    public Comment replaceComment(@RequestBody Comment newComment, @PathVariable Long id) {
-        return repository.findById(id).map(comment -> {
-            comment.setContent(newComment.getContent());
+    public ResponseEntity<?> replaceComment(@RequestBody Comment newComment, @PathVariable Long id) {
+        Comment updatedComment = repository.findById(id).map(comment -> {
             comment.setTitle(newComment.getTitle());
-            return repository.save(newComment);
+            comment.setContent(newComment.getContent());
+            return repository.save(comment);
         }).orElseGet(() -> {
             newComment.setComment_id(id);
             return repository.save(newComment);
         });
+        EntityModel<Comment> entityModel = assembler.toModel(updatedComment);
+        return ResponseEntity.created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()).body(entityModel);
     }
 
     @DeleteMapping("/comments/{id}")
-    public void deleteComment(@PathVariable Long id) {
-        repository.deleteById(id);
+    public ResponseEntity<?> deleteComment(@PathVariable Long id) {
+        Optional<Comment> commentOptional = repository.findById(id);
+        if(commentOptional.isPresent()) {
+            repository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+
     }
 }
